requirements.txt
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
PyPDF2==3.0.1
pytesseract==0.3.13
